Move Boards+BoardsAction+BoardsReducer to entities/board
