+++ src/entities/board/BoardsReducer.js
